requirements.txt
certifi==2020.6.20
chardet==3.0.4
dateparser==0.7.6
idna==2.10
influxdb==5.3.0
msgpack==0.6.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.15
requests==2.24.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
